src/requirements.txt
agency_swarm==0.4.4
fastapi==0.115.6
jsonschema==4.23.0
pandas==2.2.3
psycopg2_binary==2.9.10
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
tweepy==4.14.0
uvicorn==0.34.0
aiohttp==3.10.3
websockets==10.4
mem0ai==0.1.67
